+++ R/is_model_supported.R
@@ -146,6 +146,8 @@ supported_models <- function() {
     "mle2",
     "mlm",
     "mclogit",
+    "meta_fixed",
+    "meta_random",
     "mlogit",
     "mmlogit",
     "multinom",