+++ R/model_info.R
@@ -238,6 +238,12 @@ model_info.aovlist <- model_info.mmclogit
 model_info.rma <- model_info.mmclogit
 
 #' @export
+model_info.meta_random <- model_info.mmclogit
+
+#' @export
+model_info.meta_fixed <- model_info.mmclogit
+
+#' @export
 model_info.metaplus <- model_info.mmclogit
 
 #' @export